Added a few tests and fixed issues they unearthed
